+++ getx_route_annotations/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.1.0
+
+- support generate binding for GetPage use Annotation.