+++ getx_route_generator/CHANGELOG.md
@@ -10,3 +10,11 @@
 ## 1.0.2
 
 - update the README.md
+
+## 1.1.0
+
+- support generate binding for GetPage use Annotation.
+
+## 1.1.1
+
+- update the version of dependencies.